The benefit of iterating over a list is that we avoid having to sort results. With pre-sorted results we can stop searching once we've reached the requested limit.
On the downside, if the sorted list is much larger than the smallest set, you're
likely to have a ton of misses. In such cases, it makes more sense to iterate
over the set and sort the full final result. For example, given a set of 50 items
and a list of 100000, the difference between the two is about 20us vs 20ms.

This is an initial implementation. It kicks-in when the set is < 100 items
and the list is > 1000. Need to possibly adjust this. Also, the code isn't
particularly elegant in this first iteration.